refactor: move Blake3Cid to main for now
