+++ Cargo.toml
@@ -18,12 +18,12 @@ blake3 = "1.3.3"
 bytes = "1"
 clap = { version = "4", features = ["derive"], optional = true }
 console = { version = "0.15.5", optional = true }
-data-encoding = "2.3.3"
+data-encoding = { version = "2.3.3", optional = true }
 der = { version = "0.6", features = ["alloc", "derive"] }
 ed25519-dalek = { version = "1.0.1", features = ["serde"] }
 futures = "0.3.25"
 indicatif = { version = "0.17", features = ["tokio"], optional = true }
-multibase = "0.9.1"
+multibase = { version = "0.9.1", optional = true }
 postcard = { version = "1", default-features = false, features = ["alloc", "use-std", "experimental-derive"] }
 quinn = "0.9.3"
 rand = "0.7"
@@ -50,7 +50,7 @@ testdir = "0.7.1"
 
 [features]
 default = ["cli"]
-cli = ["clap", "console", "indicatif"]
+cli = ["clap", "console", "indicatif", "data-encoding", "multibase"]
 
 [[bin]]
 name = "sendme"
